Splitting CSV rows into XML files per row
Using Pandas, the CSV is loaded in as a data frame. One column's content is replaced with Regex and a new data frame is created. The XML structure is then declared and formatted. Finally the loop iterates through the total number of rows in the CSV (length) and creates a new XML with the specified structure and data for that loop. Each XML file created uses the ID column from the dataframe as file name.